Nit changes to HomeActivity/Fragment
