+++ app/src/main/java/org/oppia/app/home/HomeActivity.kt
@@ -6,11 +6,11 @@ import javax.inject.Inject
 
 /** The central activity for all users entering the app. */
 class HomeActivity : InjectableAppCompatActivity() {
-  @Inject lateinit var homeActivityController: HomeActivityController
+  @Inject lateinit var homeActivityPresenter: HomeActivityPresenter
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     activityComponent.inject(this)
-    homeActivityController.handleOnCreate()
+    homeActivityPresenter.handleOnCreate()
   }
 }
