+++ app/src/main/java/org/oppia/app/home/HomeActivityPresenter.kt
@@ -1,0 +1,24 @@
+package org.oppia.app.home
+
+import androidx.appcompat.app.AppCompatActivity
+import org.oppia.app.R
+import org.oppia.app.activity.ActivityScope
+import javax.inject.Inject
+
+/** The controller for [HomeActivity]. */
+@ActivityScope
+class HomeActivityPresenter @Inject constructor(private val activity: AppCompatActivity) {
+  fun handleOnCreate() {
+    activity.setContentView(R.layout.home_activity)
+    if (getHomeFragment() == null) {
+      activity.supportFragmentManager.beginTransaction().add(
+        R.id.home_fragment_placeholder,
+        HomeFragment()
+      ).commitNow()
+    }
+  }
+
+  private fun getHomeFragment(): HomeFragment? {
+    return activity.supportFragmentManager.findFragmentById(R.id.home_fragment_placeholder) as HomeFragment?
+  }
+}
